FEAT: deletekey

+++ src/routes/api/users/deleteAccount.ts
@@ -2,6 +2,9 @@ import express, { Request, Response } from "express";
 import { Pool } from "../../../../config/connectDB";
 import { PoolConnection,  ResultSetHeader } from "mysql2/promise";
 import { DeleteAccountReqeust, DeleteAccountResponse } from "../../../types/user";
+import config from 'config';
+
+const managerKey = config.get<string>("managerKey");
 
 const router = express.Router();
 
@@ -11,7 +14,18 @@ const router = express.Router();
 router.delete("/:user_id", async (req: Request<DeleteAccountReqeust['params']>, res: Response) => {
   const connection: PoolConnection = await Pool.getConnection();
   const { user_id } = req.params;
+  const { deleteKey } : DeleteAccountReqeust['body'] = req.body;
   console.log(user_id)
+
+  if (!managerKey.includes(deleteKey)) {
+    const response = {
+      statusCode: 403,
+      message: "Invalid invitation key. Access denied."
+    }
+    console.log(response);
+    return res.status(403).json(response);
+  }
+
   try {
     await connection.execute<ResultSetHeader>(
       `DELETE FROM hobit.users WHERE id = ?`,
